refactor: удалить закомментированные и неиспользуемые строки
