+++ statistics/stat-client/src/main/java/ru/practicum/ewm/controller/HitController.java
@@ -2,16 +2,13 @@ package ru.practicum.ewm.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import ru.practicum.ewm.model.Hit;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewm.model.HitClient;
+import ru.practicum.ewm.model.Hit;
+import ru.practicum.ewm.service.HitClient;
 
 import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 
 @RestController
 @RequiredArgsConstructor
