--- statistics/stat-client/src/main/java/ru/practicum/ewm/exception/ErrorHandler.java
@@ -1,16 +1,0 @@
-package ru.practicum.ewm.exception;
-
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ErrorHandler {
-
-/*    @ExceptionHandler
-    @ResponseStatus(HttpStatus.CONFLICT)
-    public ErrorResponse handlePostNotFoundException(final UserAlreadyExistException e) {
-        return new ErrorResponse(
-                e.getMessage()
-        );
-    }*/
-}
