--- statistics/stat-client/src/main/java/ru/practicum/ewm/model/ErrorResponse.java
@@ -1,4 +1,0 @@
-package ru.practicum.ewm.model;
-
-public class ErrorResponse {
-}
