+++ statistics/stat-dto/src/main/java/ru/practicum/ewm/model/Hit.java
@@ -8,7 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-
 import java.time.LocalDateTime;
 
 @Data
