+++ statistics/stat-service/src/main/java/ru/practicum/ewm/repository/HitRepository.java
@@ -1,11 +1,10 @@
 package ru.practicum.ewm.repository;
 
-import ru.practicum.ewm.model.Hit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import ru.practicum.ewm.model.Hit;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 public interface HitRepository extends JpaRepository<Hit, Long> {
 
@@ -22,19 +21,6 @@ public interface HitRepository extends JpaRepository<Hit, Long> {
             "AND h.timestamp < ?2 " +
             "AND h.uri = ?3")
     int countHitsWithUriAndNoUniqueIps(LocalDateTime start, LocalDateTime end, String uri);
-
-    @Query("SELECT COUNT(*) " +
-            "FROM Hit AS h " +
-            "WHERE h.timestamp > ?1 " +
-            "AND h.timestamp < ?2 " +
-            "GROUP BY h.ip")
-    int countHitsWithUniqueIpsAndNoUris(LocalDateTime start, LocalDateTime end);
-
-    @Query("SELECT COUNT(*) " +
-            "FROM Hit AS h " +
-            "WHERE h.timestamp > ?1 " +
-            "AND h.timestamp < ?2")
-    int countHitsWithNoUniqueIpsAndNoUris(LocalDateTime start, LocalDateTime end);
 
     @Query("SELECT h.app " +
             "FROM Hit AS h " +
