+++ statistics/stat-service/src/main/java/ru/practicum/ewm/service/HitService.java
@@ -3,7 +3,6 @@ package ru.practicum.ewm.service;
 import ru.practicum.ewm.dto.Stats;
 import ru.practicum.ewm.model.Hit;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 public interface HitService {
